\name{plot.tune}
\alias{plot.tune}
\title{Plot performance of tuning object}
\description{
  Contour or perspective plot to visualize the performace of tuning results. 
  It is limited to two tuning parameters.
}
\usage{
\method{plot}{tune}(x, type = c("contour", "perspective"), theta = 60,
          col = "lightblue", main = NULL, xlab = NULL, ylab = NULL,
          swapxy = FALSE, transform.x = NULL, transform.y = NULL,
          transform.z = NULL, color.palette = hsv_palette(),
          nlevels = 20, ...)

}
\arguments{
  \item{x}{an object of class \code{tune} generated by the \code{\link{tune.svm}}}
  \item{type}{string indicating whether a contour plot or a perspective plot is used if two parameters shall be plotted. Ignored if only one tuning parameter in the \code{x}.}
  \item{theta}{numeric value indicating angle of azimuthal direction for perspective plot.}
  \item{col}{the color(s) of the surface facets for the perspective plot. Ignored if transparent colors are specified.}
  \item{main}{string indicating main title.}
  \item{xlab}{string or expression indicating titles for the X axes.}
  \item{ylab}{string or expression indicating titles for the Y axes.}
  \item{swapxy}{logical value indicating whether the parameter axes are swapped, only used in case of two parameters.}
  \item{transform.x}{function indicating how to transform the first parameter in X axis. Ignored if \code{NULL}.}
  \item{transform.y}{function indicating how to transform the second parameter in Y axis. Ignored if \code{NULL}.}
  \item{transform.z}{function indicating how to transform the error measures in Z axis in the perspective plot or color palette in the contour plot. Ignored if \code{NULL}.}
  \item{color.palette}{color palette used in the contour plot.}
  \item{nlevels}{integer value indicating number of levels used in the contour plot.}
  \item{\dots}{additional graphics parameters passed to \code{\link{filled.contour}} or \code{\link{persp}}.}
}
\author{
  David Meyer (based on C/C++-code by Chih-Chung Chang and Chih-Jen Lin)\cr
  \email{David.Meyer@R-project.org}
}
\seealso{\code{\link{tune.svm}}}

\examples{
  data(iris)
  obj <- tune.svm(Species~., data = iris, sampling = "fix",
                  gamma = 2^c(-8,-4,0,4), cost = 2^c(-8,-4,-2,0))
  plot(obj, transform.x = log2, transform.y = log2)
  plot(obj, type = "perspective", theta = 120, phi = 45)
}
\keyword{plot}
\keyword{tuning}
